create basic welcome page
